Separate some properties of employee to abstract employee
