+++ composer.json
@@ -24,7 +24,7 @@
             "email": "roman.varkuta@gmail.com"
         }
     ],
-    "minimum-stability": "stable",
+    "minimum-stability": "dev",
     "autoload": {
         "psr-4": {
             "Wearesho\\Phonet\\": "src/"
